[chg] Update the function in serializers
